requirements.txt
matplotlib >= 3.0
numpy >= 1.17.0
opencv-python >= 4.0.0.21
pandas >= 1.1.0
pypng >= 0.0.16
pytorch-lightning >= 1.1.0
requests >= 2.0.0
scipy >= 1.0.0
torch >= 1.7.0
torchmetrics >= 0.2
torchvision >= 0.8.0
tqdm >= 4.41.0